agent-container/requirements.txt
requests==2.31.0
PyJWT==2.8.0
GitPython==3.1.40
python-dotenv==1.0.0